do not make multiple files when doing setup

+++ setup.sh
@@ -6,9 +6,9 @@ fi
 
 function install_package(){
     PKG_OK=$(dpkg-query -W --showformat='${Status}\n' $1|grep "install ok installed")
-    echo Checking for $1: $PKG_OK
+    echo Checking if $1 is installed: $PKG_OK
     if [ "" = "$PKG_OK" ]; then
-        echo "No $1. Setting up $1."
+        echo "$1 not installed. Setting up $1."
         sudo apt-get --yes install $1
     fi
 }
@@ -24,11 +24,15 @@ python3 -m pip install ffmpeg-python
 python3 -m pip install pyqt5
 
 echo "Setting up video loopback"
-echo options v4l2loopback devices=1 video_nr=20 \
-card_label="X920" exclusive_caps=1 | sudo tee -a \
-/etc/modprobe.d/virtualcam.conf
+if [ ! -f /etc/modprobe.d/virtualcam.conf ]; then
+    echo options v4l2loopback devices=1 video_nr=20 \
+    card_label="X920" exclusive_caps=1 | sudo tee -a \
+    /etc/modprobe.d/virtualcam.conf
+fi
 
-echo v4l2loopback | sudo tee -a /etc/modules-load.d/virtualcam.conf
+if [ ! -f /etc/modules-load.d/virtualcam.conf ]; then
+    echo v4l2loopback | sudo tee -a /etc/modules-load.d/virtualcam.conf
+fi
 
 sudo modprobe -r v4l2loopback
 sudo modprobe v4l2loopback
